Incorporates PR comments. Moves core tests into one file
